Added feature to keep in DataStore.

+++ src/Analyzer/code/java/metrics/indicators/CodeLines.java
@@ -2,17 +2,17 @@ package Analyzer.code.java.metrics.indicators;
 
 public class CodeLines {
 
-    private Integer blankLines;
-    private Integer commentLines;
-    private Integer effectiveLines;
+    private Double blankLines;
+    private Double commentLines;
+    private Double effectiveLines;
 
     public CodeLines() {
-        this.blankLines = 0;
-        this.commentLines = 0;
-        this.effectiveLines = 0;
+        this.blankLines = 0.0;
+        this.commentLines = 0.0;
+        this.effectiveLines = 0.0;
     }
 
-    public Integer getBlankLines() {
+    public Double getBlankLines() {
         return blankLines;
     }
 
@@ -20,7 +20,7 @@ public class CodeLines {
         this.blankLines++;
     }
 
-    public Integer getCommentLines() {
+    public Double getCommentLines() {
         return commentLines;
     }
 
@@ -28,7 +28,7 @@ public class CodeLines {
         this.commentLines++;
     }
 
-    public Integer getEffectiveLines() {
+    public Double getEffectiveLines() {
         return effectiveLines;
     }
 
@@ -36,7 +36,7 @@ public class CodeLines {
         this.effectiveLines++;
     }
 
-    public Integer getAllLines() {
+    public Double getAllLines() {
         return effectiveLines + blankLines + commentLines;
     }
